Sentry: Move import to app.module
