+++ package.json
@@ -67,7 +67,7 @@
     "@nguniversal/module-map-ngfactory-loader": "^7.1.0",
     "@ngx-translate/core": "^10.0.2",
     "@ngx-translate/http-loader": "^3.0.1",
-    "@sentry/browser": "^5.0.8",
+    "@sentry/browser": "^5.1.0",
     "@sentry/node": "^4.5.4",
     "angular-sortablejs": "^2.6.0",
     "angular2-medium-editor": "1.2.0",
