+++ src/app/environment.ts
@@ -2,10 +2,6 @@
 // rc2 workaround
 import { enableDebugTools, disableDebugTools } from '@angular/platform-browser';
 import { ApplicationRef, enableProdMode, ErrorHandler, Injectable } from '@angular/core';
-// Error tracking utility for sentry.io
-import * as Sentry from '@sentry/browser';
-// Project Dependencies
-import { version } from '../../package.json'
 // Env vars
 import { environment } from 'environments/environment'
 
@@ -18,20 +14,6 @@ let PROVIDERS: any[] = [
 // https://github.com/angular/angular/blob/86405345b781a9dc2438c0fbe3e9409245647019/TOOLS_JS.md
 let _decorateModuleRef = function identity<T>(value: T): T { return value; };
 
-// Sentry Raven reporter
-Sentry.init({ 
-  dsn: 'https://9ef1f98534914bf6826e202370d1f627@sentry.io/209953',
-  release: version
-});
-@Injectable()
-export class SentryErrorHandler implements ErrorHandler {
-  constructor() {}
-  handleError(error) {
-    const eventId = Sentry.captureException(error.originalError || error);
-    Sentry.showReportDialog({ eventId });
-  }
-}
-
 if (environment.production) {
   // Production
   disableDebugTools();
@@ -39,8 +21,6 @@ if (environment.production) {
 
   PROVIDERS = [
     ...PROVIDERS,
-    // Use Sentry error handler
-    { provide: ErrorHandler, useClass: SentryErrorHandler }
     // custom providers in production
   ];
